+++ CHANGELOG.md
@@ -13,7 +13,9 @@
    * Add Admin Info
 * App Detail
    * Subscribe Button - wrong state
-
+* Home, Use cases and Company roles
+   * Mobile friendly header and stage navigation
+   * Slider stage header
 * Bugfix
    * Service Release Process
       * Service Release process not working